major refactoring
completely change the structure of the project.
This is not complete at all, project is completely broken right now.
