+++ package.json
@@ -14,7 +14,9 @@
     "grunt-traceur": "~0.1.2",
     "grunt-contrib-copy": "~0.5.0",
     "grunt-contrib-watch": "~0.6.1",
-    "grunt-exec": "~0.4.5"
+    "grunt-exec": "~0.4.5",
+    "gulp": "~3.6.2",
+    "gulp-nodemon": "~1.0.3"
   },
   "dependencies": {
     "traceur": "0.0.42",
